backend/pyproject.toml
[tool.poetry]
name = "mychat_backend"
description = "RAG backend using FastAPI, Brave Search API, and Ollama"
authors = ["Giuseppe Marco Randayyo"]
readme = "README.md"
version = "1.0.0"

[tool.poetry.dependencies]
python = ">=3.10, <3.13"
fastapi = "^0.111.0"
uvicorn = {extras = ["standard"], version = "^0.30.0"}
python-dotenv = "^1.0.1"
requests = "^2.32.0"
httpx = "^0.27.0"
pydantic = "^2.8.2" 


[poetry.group.dev.dependencies]
pre-commit = "^2.21.0"
flake8 = "^5.0.4"
pytest = "^7.1.3"
pytest-cov = "^4.0.0"
coverage = "^6.4.4"
dlint = "^0.13.0"
flake8-2020 = "^1.7.0"
flake8-bugbear = "^22.8.23"
flake8-blind-except = "^0.2.1"
flake8-builtins = "^1.5.3"
flake8-class-newline = "^1.6.0"
flake8-comprehensions = "^3.10.0"
flake8-deprecated = "^1.3"
flake8-import-order = "^0.18.2"
flake8-quotes = "^3.3.1"
pytest-repeat = "^0.9.1"
pytest-rerunfailures = "^10.3"

[tool.poetry.extras]
testing = ["pytest", "pytest-cov", "pytest-mock", "pytest-xdist"]
docs = ["sphinx", "sphinx-autodoc-typehints", "sphinx-mdinclude", "furo"]
user = []

[tool.pytest.ini_options]
markers = [
    "unit: Unit tests",
    "integration: Integration tests"
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"